Refactor the jwt service and util into system/jwt
