+++ backend/.mockery.yml
@@ -85,20 +85,20 @@ packages:
       pkgname: certmock
       filename: "{{.InterfaceName}}_mock.go"
 
+  github.com/asgardeo/thunder/internal/system/jwt:
+    config:
+      all: true
+      dir: tests/mocks/jwtmock
+      structname: '{{.InterfaceName}}Mock'
+      pkgname: jwtmock
+      filename: "{{.InterfaceName}}_mock.go"
+
   github.com/asgardeo/thunder/internal/oauth/jwks:
     config:
       all: true
       dir: tests/mocks/oauth/jwksmock
       structname: '{{.InterfaceName}}Mock'
       pkgname: jwksmock
-      filename: "{{.InterfaceName}}_mock.go"
-
-  github.com/asgardeo/thunder/internal/oauth/jwt:
-    config:
-      all: true
-      dir: tests/mocks/oauth/jwtmock
-      structname: '{{.InterfaceName}}Mock'
-      pkgname: jwtmock
       filename: "{{.InterfaceName}}_mock.go"
 
   github.com/asgardeo/thunder/internal/oauth/scope/provider:
