+++ backend/internal/executor/googleauth/googleauthexecutor.go
@@ -27,7 +27,7 @@ import (
 	"github.com/asgardeo/thunder/internal/executor/oidcauth"
 	flowconst "github.com/asgardeo/thunder/internal/flow/constants"
 	flowmodel "github.com/asgardeo/thunder/internal/flow/model"
-	jwtutils "github.com/asgardeo/thunder/internal/system/crypto/jwt/utils"
+	"github.com/asgardeo/thunder/internal/system/jwt"
 	"github.com/asgardeo/thunder/internal/system/log"
 )
 
@@ -164,7 +164,7 @@ func (g *GoogleOIDCAuthExecutor) ValidateIDToken(execResp *flowmodel.ExecutorRes
 	}
 
 	// Verify the id token signature.
-	signErr := jwtutils.VerifyJWTSignatureWithJWKS(idToken, g.GetJWKSEndpoint())
+	signErr := g.JWTService.VerifyJWTSignatureWithJWKS(idToken, g.GetJWKSEndpoint())
 	if signErr != nil {
 		execResp.Status = flowconst.ExecFailure
 		execResp.FailureReason = "ID token signature verification failed: " + signErr.Error()
@@ -172,7 +172,7 @@ func (g *GoogleOIDCAuthExecutor) ValidateIDToken(execResp *flowmodel.ExecutorRes
 	}
 
 	// Parse the JWT claims from the ID token.
-	claims, err := jwtutils.ParseJWTClaims(idToken)
+	claims, err := jwt.DecodeJWTPayload(idToken)
 	if err != nil {
 		execResp.Status = flowconst.ExecFailure
 		execResp.FailureReason = "Failed to parse ID token claims: " + err.Error()
