+++ backend/internal/system/jwt/service_test.go
@@ -19,13 +19,20 @@
 package jwt
 
 import (
+	"crypto"
+	"crypto/ecdsa"
+	"crypto/elliptic"
 	"crypto/rand"
 	"crypto/rsa"
+	"crypto/sha256"
 	"crypto/x509"
 	"encoding/base64"
 	"encoding/json"
 	"encoding/pem"
 	"errors"
+	"fmt"
+	"net/http"
+	"net/http/httptest"
 	"os"
 	"strings"
 	"testing"
@@ -120,22 +127,19 @@ func (suite *JWTServiceTestSuite) TestNewJWTService() {
 	assert.Implements(suite.T(), (*JWTServiceInterface)(nil), service)
 }
 
-func (suite *JWTServiceTestSuite) TestInit_Success() {
-	jwtService := &JWTService{
-		SystemCertificateService: suite.mockCertService,
-	}
-	err := jwtService.Init()
-
-	assert.NoError(suite.T(), err)
-	assert.NotNil(suite.T(), jwtService.privateKey)
-}
-
 func (suite *JWTServiceTestSuite) TestInitScenarios() {
 	testCases := []struct {
 		name           string
 		setupFunc      func() string
 		expectedErrMsg string
 	}{
+		{
+			name: "Success",
+			setupFunc: func() string {
+				return suite.testKeyPath // Use the existing valid key path
+			},
+			expectedErrMsg: "",
+		},
 		{
 			name: "ReadFileError",
 			setupFunc: func() string {
@@ -282,7 +286,7 @@ func (suite *JWTServiceTestSuite) TestInitScenarios() {
 			if tc.expectedErrMsg != "" {
 				assert.Error(t, err)
 				assert.Contains(t, err.Error(), tc.expectedErrMsg)
-			} else if tc.name == "PKCS8Key" {
+			} else if tc.name == "Success" || tc.name == "PKCS8Key" {
 				assert.NoError(t, err)
 				assert.NotNil(t, jwtService.privateKey)
 			} else if tc.name == "ReadFileError" || tc.name == "InvalidPKCS8Key" || tc.name == "InvalidPKCS1Key" {
@@ -517,4 +521,363 @@ func (suite *JWTServiceTestSuite) TestGenerateJWTScenarios() {
 			}
 		})
 	}
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignature() {
+	suite.mockCertService.On("GetCertificateKid").Return("test-kid", nil).Once()
+	validToken, _, err := suite.jwtService.GenerateJWT("test-subject", "test-audience", 3600, nil)
+	assert.NoError(suite.T(), err)
+
+	wrongKey, _ := rsa.GenerateKey(rand.Reader, 2048)
+
+	// Create tampered token
+	parts := []string{}
+	for _, part := range []string{"header", "payload", "signature"} {
+		jsonData, _ := json.Marshal(map[string]string{"tampered": part})
+		parts = append(parts, base64.RawURLEncoding.EncodeToString(jsonData))
+	}
+	tamperedToken := parts[0] + "." + parts[1] + "." + parts[2]
+
+	testCases := []struct {
+		name        string
+		token       string
+		publicKey   *rsa.PublicKey
+		expectError bool
+	}{
+		{"ValidToken", validToken, &suite.testPrivateKey.PublicKey, false},
+		{"WrongKey", validToken, &wrongKey.PublicKey, true},
+		{"InvalidToken", "invalid.token", &suite.testPrivateKey.PublicKey, true},
+		{"TamperedToken", tamperedToken, &suite.testPrivateKey.PublicKey, true},
+	}
+
+	for _, tc := range testCases {
+		suite.T().Run(tc.name, func(t *testing.T) {
+			err := suite.jwtService.VerifyJWTSignature(tc.token, tc.publicKey)
+			if tc.expectError {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignatureWithJWKS() {
+	suite.mockCertService.On("GetCertificateKid").Return("test-kid", nil).Once()
+
+	token, _, err := suite.jwtService.GenerateJWT("test-subject", "test-audience", 3600, nil)
+	assert.NoError(suite.T(), err)
+
+	testServer := suite.mockJWKSServer()
+	defer testServer.Close()
+
+	err = suite.jwtService.VerifyJWTSignatureWithJWKS(token, testServer.URL)
+	assert.NoError(suite.T(), err)
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignatureWithJWKSInvalidToken() {
+	testServer := suite.mockJWKSServer()
+	defer testServer.Close()
+
+	testCases := []struct {
+		name  string
+		token string
+	}{
+		{"EmptyToken", ""},
+		{"MalformedToken", "not.valid.jwt"},
+		{"InvalidFormat", "header.payload"},                 // Missing signature part
+		{"CorruptedHeader", "aGVhZGVyCg.payload.signature"}, // Non-decodable header
+	}
+
+	for _, tc := range testCases {
+		suite.T().Run(tc.name, func(t *testing.T) {
+			err := suite.jwtService.VerifyJWTSignatureWithJWKS(tc.token, testServer.URL)
+			assert.Error(t, err)
+		})
+	}
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignatureWithJWKSKeyIDNotFound() {
+	testServer := suite.mockJWKSServer()
+	defer testServer.Close()
+
+	nonExistentKidJWT := suite.createJWTWithCustomHeader(map[string]interface{}{
+		"alg": "RS256",
+		"typ": "JWT",
+		"kid": "non-existent-key-id",
+	})
+
+	err := suite.jwtService.VerifyJWTSignatureWithJWKS(nonExistentKidJWT, testServer.URL)
+	assert.Error(suite.T(), err)
+	assert.Contains(suite.T(), err.Error(), "no matching key found")
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignatureWithJWKSNoKeyID() {
+	testServer := suite.mockJWKSServer()
+	defer testServer.Close()
+
+	noKidJWT := suite.createJWTWithCustomHeader(map[string]interface{}{
+		"alg": "RS256",
+		"typ": "JWT",
+		// No kid field
+	})
+
+	err := suite.jwtService.VerifyJWTSignatureWithJWKS(noKidJWT, testServer.URL)
+	assert.Error(suite.T(), err)
+	assert.Contains(suite.T(), err.Error(), "missing 'kid' claim")
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignatureWithJWKSHTTPErrors() {
+	testCases := []struct {
+		name          string
+		setupServer   func() *httptest.Server
+		expectedError string
+	}{
+		{
+			name: "HTTPError404",
+			setupServer: func() *httptest.Server {
+				return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+					w.WriteHeader(http.StatusNotFound)
+				}))
+			},
+			expectedError: "failed to fetch JWKS, status code: 404",
+		},
+		{
+			name: "InvalidJSONResponse",
+			setupServer: func() *httptest.Server {
+				return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+					w.Header().Set("Content-Type", "application/json")
+					w.WriteHeader(http.StatusOK)
+					if _, err := w.Write([]byte("invalid json")); err != nil {
+						suite.T().Errorf("Failed to write response: %v", err)
+					}
+				}))
+			},
+			expectedError: "failed to parse JWKS",
+		},
+		{
+			name: "JWKSWithoutMatchingKid",
+			setupServer: func() *httptest.Server {
+				// Create JWKS with different kid
+				jwks := map[string]interface{}{
+					"keys": []interface{}{
+						map[string]interface{}{
+							"kty": "RSA",
+							"kid": "different-kid",
+							"n":   "some-n",
+							"e":   "AQAB",
+						},
+					},
+				}
+				jwksData, _ := json.Marshal(jwks)
+				return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+					w.Header().Set("Content-Type", "application/json")
+					w.WriteHeader(http.StatusOK)
+					if _, err := w.Write(jwksData); err != nil {
+						suite.T().Errorf("Failed to write response: %v", err)
+					}
+				}))
+			},
+			expectedError: "no matching key found",
+		},
+		{
+			name: "InvalidJWKFormat",
+			setupServer: func() *httptest.Server {
+				// Create JWKS with invalid JWK (missing n and e)
+				jwks := map[string]interface{}{
+					"keys": []interface{}{
+						map[string]interface{}{
+							"kty": "RSA",
+							"kid": "test-kid",
+							// Missing n and e
+						},
+					},
+				}
+				jwksData, _ := json.Marshal(jwks)
+				return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+					w.Header().Set("Content-Type", "application/json")
+					w.WriteHeader(http.StatusOK)
+					if _, err := w.Write(jwksData); err != nil {
+						suite.T().Errorf("Failed to write response: %v", err)
+					}
+				}))
+			},
+			expectedError: "failed to convert JWK to RSA public key",
+		},
+		{
+			name: "InvalidTokenSignature",
+			setupServer: func() *httptest.Server {
+				return suite.mockJWKSServer()
+			},
+			expectedError: "invalid token signature",
+		},
+	}
+
+	for _, tc := range testCases {
+		suite.T().Run(tc.name, func(t *testing.T) {
+			testServer := tc.setupServer()
+			defer testServer.Close()
+
+			var token string
+			if tc.name == "InvalidTokenSignature" {
+				// Create a token with wrong signature
+				token = suite.createJWTWithCustomHeader(map[string]interface{}{
+					"alg": "RS256",
+					"typ": "JWT",
+					"kid": "test-kid",
+				})
+				// Modify the last part (signature) to make it invalid
+				parts := strings.Split(token, ".")
+				parts[2] = "invalid-signature"
+				token = strings.Join(parts, ".")
+			} else {
+				suite.mockCertService.On("GetCertificateKid").Return("test-kid", nil).Once()
+				token, _, _ = suite.jwtService.GenerateJWT("test-subject", "test-audience", 3600, nil)
+			}
+
+			err := suite.jwtService.VerifyJWTSignatureWithJWKS(token, testServer.URL)
+			assert.Error(t, err)
+			assert.Contains(t, err.Error(), tc.expectedError)
+		})
+	}
+}
+
+func (suite *JWTServiceTestSuite) TestVerifyJWTSignatureWithJWKSNetworkError() {
+	// Test with invalid URL to trigger network error
+	suite.mockCertService.On("GetCertificateKid").Return("test-kid", nil).Once()
+	token, _, err := suite.jwtService.GenerateJWT("test-subject", "test-audience", 3600, nil)
+	assert.NoError(suite.T(), err)
+
+	err = suite.jwtService.VerifyJWTSignatureWithJWKS(token, "http://localhost:99999/invalid")
+	assert.Error(suite.T(), err)
+	assert.Contains(suite.T(), err.Error(), "failed to fetch JWKS")
+}
+
+func (suite *JWTServiceTestSuite) TestInitErrorConditions() {
+	testCases := []struct {
+		name           string
+		setupFunc      func() string
+		expectedErrMsg string
+	}{
+		{
+			name: "PKCS8NonRSAKey",
+			setupFunc: func() string {
+				// Create an ECDSA private key (non-RSA)
+				privateKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+				assert.NoError(suite.T(), err)
+
+				pkcs8Bytes, err := x509.MarshalPKCS8PrivateKey(privateKey)
+				assert.NoError(suite.T(), err)
+
+				pkcs8KeyPEM := pem.EncodeToMemory(&pem.Block{
+					Type:  "PRIVATE KEY",
+					Bytes: pkcs8Bytes,
+				})
+
+				tempFile, err := os.CreateTemp("", "ecdsa_key_*.pem")
+				assert.NoError(suite.T(), err)
+				suite.tempFiles = append(suite.tempFiles, tempFile.Name())
+
+				_, err = tempFile.Write(pkcs8KeyPEM)
+				assert.NoError(suite.T(), err)
+				err = tempFile.Close()
+				assert.NoError(suite.T(), err)
+
+				return tempFile.Name()
+			},
+			expectedErrMsg: "not an RSA private key",
+		},
+	}
+
+	for _, tc := range testCases {
+		suite.T().Run(tc.name, func(t *testing.T) {
+			jwtService := &JWTService{
+				SystemCertificateService: suite.mockCertService,
+			}
+
+			thunderRuntime := config.GetThunderRuntime()
+			originalKeyFile := thunderRuntime.Config.Security.KeyFile
+			thunderRuntime.Config.Security.KeyFile = tc.setupFunc()
+
+			err := jwtService.Init()
+			assert.Error(t, err)
+			assert.Contains(t, err.Error(), tc.expectedErrMsg)
+
+			thunderRuntime.Config.Security.KeyFile = originalKeyFile
+		})
+	}
+}
+
+// Helper method to create a JWT with a custom header
+func (suite *JWTServiceTestSuite) createJWTWithCustomHeader(header map[string]interface{}) string {
+	// Create payload
+	payload := map[string]interface{}{
+		"sub":  "1234567890",
+		"name": "Test User",
+		"iat":  time.Now().Unix(),
+		"exp":  time.Now().Add(time.Hour).Unix(),
+	}
+
+	headerJSON, _ := json.Marshal(header)
+	payloadJSON, _ := json.Marshal(payload)
+
+	// Encode header and payload
+	headerBase64 := base64.RawURLEncoding.EncodeToString(headerJSON)
+	payloadBase64 := base64.RawURLEncoding.EncodeToString(payloadJSON)
+
+	// Create signature input
+	signingInput := headerBase64 + "." + payloadBase64
+
+	// Sign
+	hashed := sha256.Sum256([]byte(signingInput))
+	signature, err := rsa.SignPKCS1v15(rand.Reader, suite.testPrivateKey, crypto.SHA256, hashed[:])
+	if err != nil {
+		suite.T().Fatalf("Failed to sign JWT: %v", err)
+	}
+
+	// Encode signature
+	signatureBase64 := base64.RawURLEncoding.EncodeToString(signature)
+
+	// Create full JWT
+	return headerBase64 + "." + payloadBase64 + "." + signatureBase64
+}
+
+// Helper method to create mock JWKS data
+func (suite *JWTServiceTestSuite) createMockJWKSData() string {
+	n := base64.RawURLEncoding.EncodeToString(suite.testPrivateKey.PublicKey.N.Bytes())
+
+	// Convert exponent to bytes
+	eBytes := []byte{1, 0, 1} // 65537 in big-endian
+	e := base64.RawURLEncoding.EncodeToString(eBytes)
+
+	jwk := map[string]interface{}{
+		"kty": "RSA",
+		"n":   n,
+		"e":   e,
+		"kid": "test-kid",
+		"use": "sig",
+		"alg": "RS256",
+	}
+
+	jwks := map[string]interface{}{
+		"keys": []interface{}{jwk},
+	}
+
+	jwksData, _ := json.Marshal(jwks)
+	return string(jwksData)
+}
+
+// Helper method to mock a JWKS server
+func (suite *JWTServiceTestSuite) mockJWKSServer() *httptest.Server {
+	jwksData := suite.createMockJWKSData()
+
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		if _, err := fmt.Fprintln(w, jwksData); err != nil {
+			suite.T().Errorf("Failed to write JWKS response: %v", err)
+		}
+	}))
+
+	return server
 }
