+++ backend/internal/system/services/tokenintrospectservice.go
@@ -21,8 +21,8 @@ package services
 import (
 	"net/http"
 
-	"github.com/asgardeo/thunder/internal/oauth/jwt"
 	"github.com/asgardeo/thunder/internal/oauth/oauth2/introspect"
+	"github.com/asgardeo/thunder/internal/system/jwt"
 	"github.com/asgardeo/thunder/internal/system/server"
 )
